+++ src/cc65/symtab.c
@@ -671,7 +671,7 @@ DefOrRef* AddDefOrRef (SymEntry* E, unsigned Flags)
     DOR = xmalloc (sizeof (DefOrRef));
     CollAppend (E->V.L.DefsOrRefs, DOR);
     DOR->Line = GetCurrentLine ();
-    DOR->LocalsBlockId = (ssize_t)CollLast (&CurrentFunc->LocalsBlockStack);
+    DOR->LocalsBlockId = CollLast (&CurrentFunc->LocalsBlockStack);
     DOR->Flags = Flags;
     DOR->StackPtr = StackPtr;
     DOR->Depth = CollCount (&CurrentFunc->LocalsBlockStack);
@@ -729,9 +729,9 @@ SymEntry* AddLabelSym (const char* Name, unsigned Flags)
                 /* Are we jumping into a block with initalization of an object that
                 ** has automatic storage duration? Let's emit a warning.
                 */
-                if ((size_t)CollLast (AIC) != (size_t)(DOR->LocalsBlockId) &&
+                if ((long)CollLast (AIC) != DOR->LocalsBlockId &&
                     (CollCount (AIC) < DOR->Depth ||
-                    (size_t)CollAt (AIC, DOR->Depth - 1) != DOR->LocalsBlockId)) {
+                    (long)CollAt (AIC, DOR->Depth - 1) != DOR->LocalsBlockId)) {
                     Warning ("Goto at line %d to label %s jumps into a block with "
                     "initialization of an object that has automatic storage duration",
                     GetCurrentLine (), Name);
@@ -758,9 +758,9 @@ SymEntry* AddLabelSym (const char* Name, unsigned Flags)
                 /* Are we jumping into a block with initalization of an object that
                 ** has automatic storage duration? Let's emit a warning.
                 */
-                if ((ssize_t)CollLast (AIC) != DOR->LocalsBlockId &&
+                if ((long)CollLast (AIC) != DOR->LocalsBlockId &&
                     (CollCount (AIC) >= DOR->Depth ||
-                    (ssize_t)CollLast (AIC) >= (long)DOR->Line))
+                    (long)CollLast (AIC) >= (long)DOR->Line))
                     Warning ("Goto at line %d to label %s jumps into a block with "
                     "initialization of an object that has automatic storage duration",
                     DOR->Line, Name);
